pyproject.toml
[project]
name = "vcs-gateway"
version = "0.1.0"
description = "VCS Gateway — Developer Growth Intelligence Platform"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "asyncpg>=0.29",
    "aio-pika>=9.4",
    "redis[hiredis]>=5.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "structlog>=24.2",
    "alembic>=1.13",
    "httpx>=0.27",
    "orjson>=3.10",
    "tenacity>=8.3",
    "uuid7>=0.1",
    # OpenTelemetry — tracing + metrics
    "opentelemetry-sdk>=1.25",
    "opentelemetry-api>=1.25",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-asyncpg>=0.46b0",
    "opentelemetry-instrumentation-aio-pika>=0.46b0",
    "opentelemetry-instrumentation-redis>=0.46b0",
    "opentelemetry-instrumentation-httpx>=0.46b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "testcontainers[postgres,rabbitmq,redis]>=4.7",
    "mypy>=1.10",
    "types-redis>=4.6",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vcs_gateway"]

# ---------------------------------------------------------------------------
# uv
# ---------------------------------------------------------------------------
[tool.uv]
dev-dependencies = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "testcontainers[postgres,rabbitmq,redis]>=4.7",
    "mypy>=1.10",
    "types-redis>=4.6",
    "ruff>=0.4",
]

# ---------------------------------------------------------------------------
# Ruff — lint + format
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.lint.isort]
known-first-party = ["vcs_gateway"]

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------------------
# Mypy — strict type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
exclude = ["migrations/"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: pure logic tests, no I/O",
    "integration: real DB/Redis/RabbitMQ via testcontainers",
    "e2e: full flow tests",
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src/vcs_gateway"]
omit = ["migrations/*", "tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
